--- a/test/meshes/nastran/cylinder_cells_first.fem
+++ b/test/meshes/nastran/cylinder_cells_first.fem
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7b6d7151d15f16f2267727e474a19388d50106465ebf26af9a1496c7a9edb404
+size 11447