- Remove the costly tell / seek for identifying 2nd order solid elements
- Nodes can now be defined after elements
- Correct node ordering for certain cells

1. This fixes #603
2. Reading efficiency is greatly improved for solid elements. For the tetrahedral mesh used for performance benchmarking (400k points and 2M tetrahedra), the time has been reduced from 1 minute to 20 s.

Before
```
CPU times: user 53.7 s, sys: 5.23 s, total: 58.9 s
Wall time: 59 s
```

After
```
CPU times: user 21 s, sys: 705 ms, total: 21.7 s
Wall time: 21.6 s
```